[refactor] 1、优化命名格式（变量名、接口名） 2、优化宏定义变量

+++ src/application/application.h
@@ -5,14 +5,13 @@
 #define APP Application::GetInstance()
 
 #include <Windows.h>
-
 #include "camera.h"
 
 class Application 
 {
 public:
-    static constexpr const LONG kDefaultWindowWidth = 800;
-    static constexpr const LONG kDefaultWindowHeight = 600;
+    static constexpr const int32_t kDefaultWindowWidth = 800;
+    static constexpr const int32_t kDefaultWindowHeight = 600;
     static constexpr const CHAR* kDefaultWindowTitle = "DragonRenderer";
     static Application* GetInstance();
 public:
@@ -22,15 +21,15 @@ public:
     //注册、创建、显示窗口
     bool InitMainWindow(HINSTANCE program_instance, 
         const CHAR* main_window_title = kDefaultWindowTitle,
-        LONG main_window_width = kDefaultWindowWidth, 
-        LONG main_window_height = kDefaultWindowHeight);
+        int32_t main_window_width = kDefaultWindowWidth, 
+        int32_t main_window_height = kDefaultWindowHeight);
 
 	void ProcessMessage(HWND window_handler, UINT message_id, WPARAM message_wparam, LPARAM message_lparam);
 
 	bool DispatchMessageLoop();
 
-    LONG GetMainWindowWidth() const { return main_window_width_; }
-    LONG GetMainWindowHeight() const { return main_window_height_; }
+    int32_t GetMainWindowWidth() const { return main_window_width_; }
+    int32_t GetMainWindowHeight() const { return main_window_height_; }
 
     void* GetRenderBuffer() const { return canvas_buffer_;}
 
@@ -53,8 +52,8 @@ private:
 
     Camera* camera_ { nullptr };
 
-    LONG main_window_width_ {0};
-    LONG main_window_height_ {0};
+    int32_t main_window_width_ {0};
+    int32_t main_window_height_ {0};
     const CHAR* main_window_title_ { nullptr };
     const WCHAR* main_window_class_name_ = L"DragonWindowClass";
